// generated/src/app/api-models/pmb-cust-fin-accts.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class PmbCustFinAccts {
  customerType: string;
  customerId: string;
  acctNo: string;
  effectiveFromDate: Date;
  acctUse: string;
  effectiveToDate: Date;
  acctType: string;
  routingNo: number;
  acctStatus: string;
  ccExpireDate: Date;
  securityCode: string;
  insertDatetime: Date;
  insertUser: string;
  insertProcess: string;
  updateDatetime: Date;
  updateUser: string;
  updateProcess: string;
}